refactor(scripts): split the residency guard by rule and share its AST walk
residency_boundary_test.go had grown to 1702 lines holding four independent
rules, their controls, the shared scaffolding and the baseline reader. Nothing
about a rule told you which of the four it belonged to except its position in
the file, and the two AST rules each carried their own copy of the same
recursive walk-and-parse — including the prune logic, which has to stay
identical across scanners or the halves police different trees.

One file per rule, plus one for the scaffolding they share and one for the
baseline they all ratchet on:

  residency_boundary_test.go          scan scope, allowlist, pattern names,
                                      the shared walk
  residency_grep_rule_test.go         rules a-d, the vocabulary census
  residency_signature_rule_test.go    ast:returns-store-list
  residency_expression_rules_test.go  ast:vocabulary-alias,
                                      ast:plan-leg-store-chain,
                                      ast:uncounted-call-spelling
  residency_halves_agree_test.go      shell and Go halves police the same tree
  residency_ratchet_test.go           the one baseline

residencyParseTree does the walk once and hands both AST rules a parsed tree;
each rule is now a pure counter over it (countStoreListSignatures,
countResidencyASTExpressions). It deliberately does not memoize across calls:
the controls falsify each rule by rewriting real files in a t.TempDir() tree
and rescanning, and a cache keyed by root would hand a control the tree from
before its own edit and report a bite that was never asked for.

The grep rule keeps its own walk. It reads lines, not syntax, and giving it a
parsed tree it would not use is a dependency for symmetry's sake.

Pure code motion: no rule changed, no pattern changed, no exemption changed.
The proof is that scripts/residency-boundary-baseline.txt does not move —
every rule counts the same sites in the same places as before.

Verified: go vet ./scripts/ clean, gofmt clean, go test ./scripts/ ok 89.5s
(the Residency rows alone, 12.3s), scripts/check-residency-boundary.sh and its
--self-test both pass, and the baseline file is untouched in this commit.
