+++ scripts/residency_halves_agree_test.go
@@ -1,0 +1,176 @@
+package scripts_test
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+	"strings"
+	"testing"
+)
+
+// The two renderings of the grep rule agree on what they police.
+//
+// scripts/check-residency-boundary.sh and residency_grep_rule_test.go read the
+// same pattern file, but each keeps its own scan scope and its own allowlist.
+// Those are holes in a guard, so they have to be identical on both sides or the
+// weaker side is the real policy — which is what these rows pin.
+
+// TestResidencyBoundaryHalvesAgree pins that the shell rendering and this one
+// police the same tree with the same exceptions.
+//
+// The strongest agreement proof is not here but in the baseline: its grep rows
+// are GENERATED by the shell half's awk pass and ASSERTED by this half's Go
+// loop, across every (file, function) key on the real tree. If the two
+// enclosing-function state machines ever disagreed about one line, one of them
+// would fail against the shared file. What is left to check is the scan scope
+// and the allowlist, which live in each half separately.
+//
+// Both are compared as EXACT SETS, parsed out of the shell's own array
+// literals rather than searched for anywhere in the file. A substring search
+// over the whole script is the wrong test twice over: it passes on a path that
+// appears only in a comment (this file's own prose names three retired
+// exemptions), and — the hole that matters — it is one-directional, so the
+// shell could exempt a file the Go half still polices and nothing would say so.
+// An exemption is a hole in a guard; it must be identical on both sides or the
+// weaker side is the real policy.
+func TestResidencyBoundaryHalvesAgree(t *testing.T) {
+	script, err := os.ReadFile(filepath.Join(residencyScriptsDir(t), "check-residency-boundary.sh"))
+	if err != nil {
+		t.Fatalf("reading the shell guard: %v", err)
+	}
+	body := string(script)
+	assertResidencySetsMatch(t, "scan dir", residencyShellArray(t, body, "scan_dirs"), residencyScanDirs)
+	goAllowed := make([]string, 0, len(residencyAllowlist))
+	for path := range residencyAllowlist {
+		goAllowed = append(goAllowed, path)
+	}
+	assertResidencySetsMatch(t, "allowlist entry", residencyShellArray(t, body, "allowlist"), goAllowed)
+	if !strings.Contains(body, "residency-boundary-patterns.txt") {
+		t.Error("the shell guard does not read the shared pattern file")
+	}
+	if !strings.Contains(body, "--self-test") {
+		t.Error("the shell guard has no --self-test; a guard nothing asserts against can be defanged silently")
+	}
+	// The pruned directories are the third exemption, and the same argument
+	// applies: a tree one half censuses and the other skips is a tree where the
+	// baseline cannot be generated by one and asserted by the other. They are
+	// pinned by name rather than as a parsed set because the shell states them in
+	// a `find -prune` expression, not an array — so this asserts the names appear
+	// in the pruning expression itself, not merely somewhere in the file.
+	prune, _, ok := strings.Cut(body, "-print0")
+	if !ok {
+		t.Fatal("the shell guard has no find ... -print0 census walk to check for prunes")
+	}
+	if _, prune, ok = strings.Cut(prune, "\tfind "); !ok {
+		t.Fatal("the shell guard's census no longer starts with find; the prune comparison is stale")
+	}
+	for _, dir := range residencyPrunedDirs {
+		if !strings.Contains(prune, "-name "+dir) {
+			t.Errorf("the shell census does not prune %q but the Go scanners do; one half walks a tree the other cannot", dir)
+		}
+	}
+}
+
+// residencyShellArray reads one `name=(...)` bash array literal out of the
+// script. It fails the test rather than returning empty on a miss: an array it
+// cannot find would silently reduce the set comparison above to "the shell
+// exempts nothing", which is the agreement failure it exists to catch.
+func residencyShellArray(t *testing.T, body, name string) []string {
+	t.Helper()
+	_, rest, ok := strings.Cut(body, "\n"+name+"=(")
+	if !ok {
+		t.Fatalf("the shell guard has no %s=( array; the halves can no longer be compared", name)
+	}
+	// The cut is on the FIRST `)`, not on a line-leading one, because the two
+	// arrays are formatted differently: scan_dirs is a single line whose `)`
+	// closes it, allowlist spans lines and closes at column 0. Both hold paths
+	// and directory names, none of which can contain a parenthesis, so the
+	// simple cut is exact for both — and a `#` field ends the scan, so a trailing
+	// comment cannot be read as an entry.
+	inner, _, ok := strings.Cut(rest, ")")
+	if !ok {
+		t.Fatalf("the shell guard's %s=( array is unterminated", name)
+	}
+	var out []string
+	for _, f := range strings.Fields(inner) {
+		if strings.HasPrefix(f, "#") {
+			break
+		}
+		out = append(out, f)
+	}
+	return out
+}
+
+func assertResidencySetsMatch(t *testing.T, what string, shell, gohalf []string) {
+	t.Helper()
+	for _, d := range residencySetDifferences(what, shell, gohalf) {
+		t.Error(d)
+	}
+}
+
+// residencySetDifferences reports each way the two sets disagree, in both
+// directions.
+//
+// It is separated from the assertion so its own controls can check the RESULT
+// rather than driving a hand-made *testing.T to see whether it failed. A
+// zero-valued T records an Errorf, but it is not running under tRunner, so the
+// day this helper gains a Fatalf the control's goroutine would simply exit and
+// the control would report something other than what it is testing.
+func residencySetDifferences(what string, shell, gohalf []string) []string {
+	in := func(xs []string, want string) bool {
+		for _, x := range xs {
+			if x == want {
+				return true
+			}
+		}
+		return false
+	}
+	var out []string
+	for _, s := range shell {
+		if !in(gohalf, s) {
+			out = append(out, fmt.Sprintf("the shell guard has %s %q and this half does not; the weaker half is the real policy", what, s))
+		}
+	}
+	for _, g := range gohalf {
+		if !in(shell, g) {
+			out = append(out, fmt.Sprintf("this half has %s %q and the shell guard does not; the weaker half is the real policy", what, g))
+		}
+	}
+	return out
+}
+
+// TestResidencyHalvesAgreementControls proves the agreement check above is not
+// fail-never. A set comparison that silently degrades to comparing nothing is
+// the classic way a two-implementation guard drifts apart unnoticed.
+func TestResidencyHalvesAgreementControls(t *testing.T) {
+	const body = "\nscan_dirs=(cmd/gc internal/api)\n\nallowlist=(\n\tcmd/gc/one.go\n)\n"
+
+	t.Run("the array parser reads both shapes", func(t *testing.T) {
+		if got := residencyShellArray(t, body, "scan_dirs"); len(got) != 2 || got[0] != "cmd/gc" || got[1] != "internal/api" {
+			t.Fatalf("single-line array: %v", got)
+		}
+		if got := residencyShellArray(t, body, "allowlist"); len(got) != 1 || got[0] != "cmd/gc/one.go" {
+			t.Fatalf("multi-line array: %v", got)
+		}
+	})
+
+	t.Run("an exemption only the shell holds is reported", func(t *testing.T) {
+		d := residencySetDifferences("allowlist entry", []string{"a.go", "b.go"}, []string{"a.go"})
+		if len(d) != 1 || !strings.Contains(d[0], `shell guard has allowlist entry "b.go"`) {
+			t.Fatalf("the shell exempting a file this half still polices was accepted: %v", d)
+		}
+	})
+
+	t.Run("an exemption only this half holds is reported", func(t *testing.T) {
+		d := residencySetDifferences("allowlist entry", []string{"a.go"}, []string{"a.go", "b.go"})
+		if len(d) != 1 || !strings.Contains(d[0], `this half has allowlist entry "b.go"`) {
+			t.Fatalf("this half exempting a file the shell still polices was accepted: %v", d)
+		}
+	})
+
+	t.Run("identical sets pass", func(t *testing.T) {
+		if d := residencySetDifferences("allowlist entry", []string{"b.go", "a.go"}, []string{"a.go", "b.go"}); len(d) != 0 {
+			t.Fatalf("equal sets in a different order must agree; the check is fail-always: %v", d)
+		}
+	})
+}
